Adding progress description to all providers
