+++ src/vs/workbench/contrib/chat/common/chatSessionsService.ts
@@ -212,6 +212,7 @@ export interface IChatSessionsService {
 	isEditable(sessionResource: URI): boolean;
 	// #endregion
 	registerModelProgressListener(model: IChatModel, callback: () => void): void;
+	getSessionDescription(chatModel: IChatModel): string | undefined;
 }
 
 export const IChatSessionsService = createDecorator<IChatSessionsService>('chatSessionsService');
